NODE-1031: Move `LogStub` to shared module.
